% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPbinary.R
\name{cov.func}
\alias{cov.func}
\title{Covariance matrix}
\usage{
cov.func(hyper, x1, x2)
}
\arguments{
\item{hyper}{the hyper parameters involved in covariance matrix}

\item{x1}{the fitst covariate}

\item{x2}{the second covariate}
}
\value{
return covariance matrix
}
\description{
A function to calculating the covariance matrix
}
