% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPbinary.R
\name{Dgamma.f}
\alias{Dgamma.f}
\title{The gradient for Laplace approximation}
\usage{
Dgamma.f(f, hyper, x1, x2, y)
}
\arguments{
\item{f}{the latent process f}

\item{hyper}{the hyper parameters involved in covariance matrix}

\item{x1}{the fitst covariate}

\item{x2}{the second covariate}

\item{y}{the response variable}
}
\value{
return the gradient
}
\description{
A function to calculating the gradient for Laplace approximation
}
