% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPbinary.R
\name{fit.GPbinary}
\alias{fit.GPbinary}
\title{The main function for fitting a binary Generalized GPR}
\usage{
fit.GPbinary(hyper, x1, x2, y)
}
\arguments{
\item{hyper}{the hyper parameters involved in covariance matrix}

\item{x1}{the fitst covariate}

\item{x2}{the second covariate}

\item{y}{the response variable}
}
\value{
a list contains hyper-parameters and the latent process f
}
\description{
A function for fitting a binary Generalized GPR and return the estimated kernel hyper-parameters
}
