% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPbinary.R
\name{GP.predict}
\alias{GP.predict}
\title{The prediction function}
\usage{
GP.predict(hyper, x1, x2, x1.new, x2.new, f.hat)
}
\arguments{
\item{hyper}{the hyper parameters obtained from fit.GPbinary()}

\item{x1}{the fitst covariate}

\item{x2}{the second covariate}

\item{x1.new}{the first new covariate}

\item{x2.new}{the second new covariate}

\item{f.hat}{the latent process obtained from fit.GPbinary()}
}
\value{
return a prediction
}
\description{
A function for predicting at a new input after we have fitted the binary generalized Gaussian process regression model with \code{fit.GPbinary()}
}
