% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPbinary.R
\name{mymatinv}
\alias{mymatinv}
\title{inverse of symmetry matrix}
\usage{
mymatinv(smatrix, sB = "sB", det = F, log = T, jitter = 1e-10)
}
\arguments{
\item{smatrix}{the symmetry matrix}

\item{sB}{sB}

\item{det}{determinant}

\item{log}{logarithm}

\item{jitter}{jiiter to aviod singular case}
}
\value{
return the inverse
}
\description{
A function to calculating the inverse of a symmetry matrix
}
