% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPbinary.R
\name{y.loglike}
\alias{y.loglike}
\title{Log-likelihood}
\usage{
y.loglike(hyper, f.hat, x1, x2, y)
}
\arguments{
\item{hyper}{the hyper parameters involved in covariance matrix}

\item{f.hat}{the latent process}

\item{x1}{the fitst covariate}

\item{x2}{the second covariate}

\item{y}{the response variable}
}
\value{
return minus log-likelihood value
}
\description{
A function to calculating the minus log-likelihood
}
